+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## [Unreleased][]
 
-[Unreleased]: https://github.com/chaostoolkit/chaostoolkit-lib/compare/1.6.0...HEAD
+[Unreleased]: https://github.com/chaostoolkit/chaostoolkit-lib/compare/1.7.0...HEAD
+
+## [1.7.0][] - 2019-09-21
+
+[1.7.0]: https://github.com/chaostoolkit/chaostoolkit-lib/compare/1.6.0...1.7.0
 
 ### Changed